% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostics.R
\name{Cooks_dist}
\alias{Cooks_dist}
\title{Cook's distance for the coefficient estimates.}
\usage{
Cooks_dist(model, cook_thr = 0, label_angle = 0)
}
\arguments{
\item{model}{An object of class "lme" representing the linear mixed-effects model fitted by \code{\link[=lmmModel]{lmmModel()}}.}

\item{cook_thr}{Threshold for the Cook's distance.}

\item{label_angle}{Angle for the label of subjects with a Cook's distance greater than \code{cook_thr}.}
}
\value{
Plot of the Cook's distance value for each subject, indicating those subjects
whose Cook's distance is greater than \code{cook_thr}.
}
\description{
Cook's distance for the coefficient estimates.
}
