% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.simulateTumorGrowth}
\alias{.simulateTumorGrowth}
\title{Helper function to simulate tumor growth data}
\usage{
.simulateTumorGrowth(
  npg = 5,
  timepoints = c(0, 3, 5, 10),
  initial_volume = 100,
  grwrControl = 0.08,
  grwrA = 0.07,
  grwrB = 0.06,
  grwrComb = 0.04,
  sd = 0.1
)
}
\arguments{
\item{npg}{Number of mouse per group.}

\item{timepoints}{Vector with the time points at which the tumor volume measurements have been performed.}

\item{initial_volume}{Initial volume for the tumor growth.}

\item{grwrControl}{Coefficient for Control treatment group tumor growth rate.}

\item{grwrA}{Coefficient for Drug A treatment group tumor growth rate.}

\item{grwrB}{Coefficient for Drug B treatment group tumor growth rate.}

\item{grwrComb}{Coefficient for Combination (Drug A + Drug B) treatment group tumor growth rate.}

\item{sd}{Standard deviation for the tumor growth.}
}
\value{
A data frame with tumor growth data in long format.
}
\description{
Helper function to simulate tumor growth data
}
