% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMM_Model.R
\name{LMM_Model}
\alias{LMM_Model}
\title{Generate Linear Mixed Model for synergy calculation}
\usage{
LMM_Model(
  data,
  Mouse,
  Day,
  Treatment,
  TV,
  C,
  A,
  B,
  AB,
  day_start = 0,
  min_obs = 1,
  plot = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame with the tumor growth data, in long format.
It must contain at least contain the following columns: mice IDs, days of follow-up (numeric number), treatment and tumor volume (numeric number).}

\item{Mouse}{String indicating the name of the column with the mice IDs.}

\item{Day}{String indicating the name of the column with the days of follow-up.}

\item{Treatment}{String indicating the name of the column with the treatment corresponding to each mouse.}

\item{TV}{String indicating the name of the column with the tumor volume.}

\item{C}{String indicating the name assigned to the 'Control' group.}

\item{A}{String indicating the name assigned to the 'Drug A' group.}

\item{B}{String indicating the name assigned to the 'Drug B' group.}

\item{AB}{String indicating the name assigned to the Combination ('Drug A' + 'Drug B') group.}

\item{day_start}{Numeric value indicating the day at which the treatment started.}

\item{min_obs}{Minimum number of observation for each mouse to be included in the analysis.}

\item{plot}{Logical indicating if a plot for the log of the relative tumor volume (RTV) vs Day for each mouse,
and the model calculated marginal slope for each treatment, should be produced.}

\item{...}{Additional arguments to be passed to \code{\link[nlme:lme]{nlme::lme()}} function.}
}
\value{
An object of class "lme" \code{\link[nlme:lme]{nlme::lme()}} representing the linear mixed-effects model fit.
}
\description{
Generate Linear Mixed Model for synergy calculation
}
