% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmSynergy.R
\name{lmmSynergy}
\alias{lmmSynergy}
\title{Synergy calculation using linear-mixed models.}
\usage{
lmmSynergy(
  model,
  method = "Bliss",
  min_time = 0,
  robustSE = FALSE,
  type = "CR2",
  ra_nsim = 1000,
  t_ci = NULL,
  show_plot = TRUE,
  ...
)
}
\arguments{
\item{model}{An object of class "lme" representing the linear mixed-effects model fitted by \code{\link[=lmmModel]{lmmModel()}}.}

\item{method}{String indicating the method for synergy calculation. Possible methods are "Bliss", "HSA" and "RA",
corresponding to Bliss, highest single agent and response additivity, respectively.}

\item{min_time}{Minimun time from which to start calculating synergy.}

\item{robustSE}{If TRUE, uncertainty is estimated using robust standard errors
using a sandwich estimate of the variance-covariance matrix of the regression coefficient estimates provided by \code{\link[clubSandwich:vcovCR]{clubSandwich::vcovCR()}}.}

\item{type}{Character string specifying which small-sample adjustment should be used, with available options "CR0", "CR1", "CR1p", "CR1S", "CR2", or "CR3".
See "Details" section of \code{\link[clubSandwich:vcovCR]{clubSandwich::vcovCR()}} for further information.}

\item{ra_nsim}{Number of simulations to calculate the synergy for Response Additivity model.}

\item{t_ci}{Time for calculation of combination index value. The value of the CI represents the proportion of tumor cell survival at time \code{t_ci}
in the drug combination group compared to the expected tumor cell survival according to the reference model. If not specified, \code{t_ci} is set to the
last time point of follow-up.}

\item{show_plot}{Logical indicating if a plot with the results of the synergy calculation should be generated.}

\item{...}{Additional arguments to be passed to \code{\link[marginaleffects:hypotheses]{marginaleffects::hypotheses()}}.}
}
\value{
The function returns a list with two elements:
\itemize{
\item \code{Constrasts}: List with the outputs of the (non)-linear test for the synergy null hypothesis obtained by \code{\link[marginaleffects:hypotheses]{marginaleffects::hypotheses()}} for each time.
See \code{\link[marginaleffects:hypotheses]{marginaleffects::hypotheses()}} for more details.
\item \code{Synergy}: Data frame with the synergy results, indicating the model of synergy ("Bliss", "HSA" or "RA"), the metric (combination index and synergy score),
the value of the metric estimate (with upper and lower confidence intervals) and the p-value, for each time.
}
}
\description{
Synergy calculation using linear-mixed models.
}
