% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmDiagnostics.R
\name{logLikSubjectDisplacements}
\alias{logLikSubjectDisplacements}
\title{Likelihood displacement for the model.}
\usage{
logLikSubjectDisplacements(
  model,
  disp_thrh = 0,
  label_angle = 0,
  var_name = NULL,
  ...
)
}
\arguments{
\item{model}{An object of class "lme" representing the linear mixed-effects model fitted by \code{\link[=lmmModel]{lmmModel()}}.}

\item{disp_thrh}{Threshold of log-likelihood displacement.}

\item{label_angle}{Angle for the label of subjects with a log-likelihood displacement greater than \code{disp_thrh}.}

\item{var_name}{Name of the variable for the weights of the model in the case that a variance structure has been specified using \code{\link[nlme:varIdent]{nlme::varIdent()}}.}

\item{...}{Extra arguments, if any, for \link[lattice:panel.xyplot]{lattice::panel.xyplot}.}
}
\value{
Returns a plot of the log-likelihood displacement values for each subject, indicating those subjects
whose contribution is greater than \code{disp_thrh}.
}
\description{
Likelihood displacement for the model.
}
