% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostics.R
\name{loo_logLik_disp}
\alias{loo_logLik_disp}
\title{Likelihood displacement for the model.}
\usage{
loo_logLik_disp(model, disp_thrh = 0, label_angle = 0, var_name = NULL, ...)
}
\arguments{
\item{model}{An object of class "lme" representing the linear mixed-effects model fitted by \code{\link[=lmm_model]{lmm_model()}}.}

\item{disp_thrh}{Threshold of log-likelihood displacement.}

\item{label_angle}{Angle for the label of subjects with a log-likelihood displacement greater than \code{disp_thrh}.}

\item{var_name}{Name of the variable for the weights of the model in the case that a variance structure has been specified using \code{\link[nlme:varIdent]{nlme::varIdent()}}.}

\item{...}{Extra arguments, if any, for \link[lattice:panel.xyplot]{lattice::panel.xyplot}. Usually passed on as
graphical parameters to low level plotting functions, or to the panel functions performing smoothing, if applicable.}
}
\value{
Returns a plot of the log-likelihood displacement values for each subject, indicating those subjects
whose contribution is greater than \code{disp_thrh}.
}
\description{
Likelihood displacement for the model.
}
