% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{Plot_exmpDt}
\alias{Plot_exmpDt}
\title{Plot of exemplary data for power calculation}
\usage{
Plot_exmpDt(
  exmpDt,
  grwrControl = NULL,
  grwrA = NULL,
  grwrB = NULL,
  grwrComb = NULL,
  sd_ranef = NULL,
  sgma = NULL
)
}
\arguments{
\item{exmpDt}{Data frame with exemplary data obtained with \code{\link[=Pwr_lmm]{Pwr_lmm()}}.}

\item{grwrControl}{Value for the label of the coefficient for Control treatment group tumor growth rate.}

\item{grwrA}{Value for the label of the coefficient for Drug A treatment group tumor growth rate.}

\item{grwrB}{Value for the label of the coefficient for Drug B treatment group tumor growth rate.}

\item{grwrComb}{Value for the label of the coefficient for Combination (Drug A + Drug B) treatment group tumor growth rate.}

\item{sd_ranef}{Value for the label of the random effects standard deviation of the model.}

\item{sgma}{Value for the label of the residuals standard deviation of the model.}
}
\value{
A ggplot2 plot (see \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} for more details) showing the regression lines corresponding
to the fixed effects for each treatment of the exemplary data for power calculations.
}
\description{
Plot of exemplary data for power calculation
}
