% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{Plot_LMM_Model}
\alias{Plot_LMM_Model}
\title{Vizualization of tumor growth data and linear mixed model fitted regression line.}
\usage{
Plot_LMM_Model(model, C, A, B, AB)
}
\arguments{
\item{model}{An object of class "lme" representing the linear mixed-effects model fitted by \code{\link[=LMM_Model]{LMM_Model()}}.}

\item{C}{String indicating the name assigned to the 'Control' group.}

\item{A}{String indicating the name assigned to the 'Drug A' group.}

\item{B}{String indicating the name assigned to the 'Drug B' group.}

\item{AB}{String indicating the name assigned to the Combination ('Drug A' + 'Drug B') group.}
}
\value{
A ggplot2 plot (see \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} for more details) showing the tumor growth data represented as log(relative tumor volume) versus time since treatment initiation.
The regression lines corresponding to the fixed effects for each treatment group are also plotted.
}
\description{
Vizualization of tumor growth data and linear mixed model fitted regression line.
}
