% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{Plot_LMM_Syn}
\alias{Plot_LMM_Syn}
\title{Visualization of synergy results obtained by \code{\link[=LMM_Syn]{LMM_Syn()}}.}
\usage{
Plot_LMM_Syn(syn_data)
}
\arguments{
\item{syn_data}{Object obtained by \code{\link[=LMM_Syn]{LMM_Syn()}} with the results of synergy calculation using linear mixed models.}
}
\value{
A ggplot2 plot (see \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} for more details) with the combination index (CI) and synergy score (SS)
estimates, confidence intervals and p-values for the synergy calculation using linear mixed models.
}
\description{
Visualization of synergy results obtained by \code{\link[=LMM_Syn]{LMM_Syn()}}.
}
