% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{Pwr_lmm}
\alias{Pwr_lmm}
\title{A priori power calculation for a hypothetical study of synergy evaluation using LMM.}
\usage{
Pwr_lmm(
  npg = 5,
  Day = c(0, 3, 5, 10),
  grwrControl,
  grwrA,
  grwrB,
  grwrComb,
  sd_ranef,
  sgma,
  sd_eval = NULL,
  sgma_eval = NULL,
  grwrComb_eval = NULL,
  method = "Bliss",
  ...
)
}
\arguments{
\item{npg}{Number of mouse per group.}

\item{Day}{Vector with the days at which the tumor volume measurements have been performed.}

\item{grwrControl}{Coefficient for Control treatment group tumor growth rate.}

\item{grwrA}{Coefficient for Drug A treatment group tumor growth rate.}

\item{grwrB}{Coefficient for Drug B treatment group tumor growth rate.}

\item{grwrComb}{Coefficient for Combination (Drug A + Drug B) treatment group tumor growth rate.}

\item{sd_ranef}{Random effects standard deviation for the model.}

\item{sgma}{Residuals standard deviation for the model.}

\item{sd_eval}{A vector with values representing the standard deviations of random effects,
through which the power for synergy calculation will be evaluated.}

\item{sgma_eval}{A vector with values representing the standard deviations of the residuals,
through which the power for synergy calculation will be evaluated.}

\item{grwrComb_eval}{A vector with values representing the coefficients for Combination treatment group tumor growth rate,
through which the power for synergy calculation will be evaluated.}

\item{method}{String indicating the method for synergy calculation. Possible methods are "Bliss", "HSA" and "RA",
corresponding to Bliss, highest single agent and response additivity, respectively.}

\item{...}{Additional parameters to be passed to \link[nlmeU:Pwr.lme]{nlmeU::Pwr.lme} method.}
}
\description{
A priori power calculation for a hypothetical study of synergy evaluation using LMM.
}
