% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmPower.R
\name{Pwr_lmm_Day}
\alias{Pwr_lmm_Day}
\title{A priori power calculation for a hypothetical study of synergy evaluation using LMM
depending on the days of follow-up or the frequency of measurements}
\usage{
Pwr_lmm_Day(
  npg = 5,
  Day = list(seq(0, 9, 3), seq(0, 21, 3), seq(0, 30, 3)),
  type = "max",
  grwrControl,
  grwrA,
  grwrB,
  grwrComb,
  sd_ranef,
  sgma,
  method = "Bliss",
  ...
)
}
\arguments{
\item{npg}{Number of mouse per group.}

\item{Day}{A list in which element is a vector with the days at which the tumor volume measurements have been performed.
If \code{type} is set to "max", each vector in the list should have the measurements taken at the same interval and differ in the final
day of follow-up. If \code{type} is set to "freq", each vector in the list should have the same final day of follow-up and
differ in the intervals at which the measurements have been taken.}

\item{type}{String indicating whether to calculate the power depending on the days of follow-up ("max"), or the frequency
of measurements ("freq").}

\item{grwrControl}{Coefficient for Control treatment group tumor growth rate.}

\item{grwrA}{Coefficient for Drug A treatment group tumor growth rate.}

\item{grwrB}{Coefficient for Drug B treatment group tumor growth rate.}

\item{grwrComb}{Coefficient for Combination (Drug A + Drug B) treatment group tumor growth rate.}

\item{sd_ranef}{Random effects standard deviation for the model.}

\item{sgma}{Residuals standard deviation for the model.}

\item{method}{String indicating the method for synergy calculation. Possible methods are "Bliss" and "HSA",
corresponding to Bliss and highest single agent, respectively.}

\item{...}{Additional parameters to be passed to \link[nlmeU:Pwr.lme]{nlmeU::Pwr.lme} method.}
}
\value{
The functions returns two plots:
\itemize{
\item A plot representing the hypothetical data, with the regression lines for each
treatment group according to \code{grwrControl}, \code{grwrA}, \code{grwrB} and \code{grwrComb} values. The values
assigned to \code{sd_ranef} and \code{sgma} are also shown.
\item A plot showing the values of the power calculation depending on the values assigned to
\code{Day}. If \code{type} is set to "max", the plot shows how the power varies depending on the maximum day of follow-up.
If \code{type} is set to "freq", the plot shows how the power varies depending on how frequently the measurements have
been performed.
}

If saved to a variable, the function returns the power for the analysis for each value specified in \code{ Day}.
}
\description{
A priori power calculation for a hypothetical study of synergy evaluation using LMM
depending on the days of follow-up or the frequency of measurements
}
