% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{Pwr_lmm_N}
\alias{Pwr_lmm_N}
\title{A priori power calculation for a hypothetical study of synergy evaluation using LMM
depending on the sample size per group}
\usage{
Pwr_lmm_N(
  npg = c(5, 8, 10),
  Day = c(0, 3, 5, 10),
  grwrControl,
  grwrA,
  grwrB,
  grwrComb,
  sd_ranef,
  sgma,
  method = "Bliss",
  ...
)
}
\arguments{
\item{npg}{A vector with the sample size (number of animals) per group to calculate the power of
the synergy analysis.}

\item{Day}{Vector with the days at which the tumor volume measurements have been performed.}

\item{grwrControl}{Coefficient for Control treatment group tumor growth rate.}

\item{grwrA}{Coefficient for Drug A treatment group tumor growth rate.}

\item{grwrB}{Coefficient for Drug B treatment group tumor growth rate.}

\item{grwrComb}{Coefficient for Combination (Drug A + Drug B) treatment group tumor growth rate.}

\item{sd_ranef}{Random effects standard deviation for the model.}

\item{sgma}{Residuals standard deviation for the model.}

\item{method}{String indicating the method for synergy calculation. Possible methods are "Bliss" and "HSA",
corresponding to Bliss and highest single agent, respectively.}

\item{...}{Additional parameters to be passed to \link[nlmeU:Pwr.lme]{nlmeU::Pwr.lme} method.}
}
\value{
The functions returns two plots:
\itemize{
\item A plot representing the hypothetical data, with the regression lines for each
treatment group according to \code{grwrControl}, \code{grwrA}, \code{grwrB} and \code{grwrComb} values. The values
assigned to \code{sd_ranef} and \code{sgma} are also shown.
\item A plot showing the values of the power calculation depending on the values assigned to
\code{npg}.
}

If save to a variable, the fuction returns the data frame with the power for the analysis for each sample size
especified in \code{npg}.
}
\description{
A priori power calculation for a hypothetical study of synergy evaluation using LMM
depending on the sample size per group
}
