% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostics.R
\name{resid_diag}
\alias{resid_diag}
\title{Diagnostics of residuals of the linear mixed model.}
\usage{
resid_diag(model, pvalue = 0.05)
}
\arguments{
\item{model}{An object of class "lme" representing the linear mixed-effects model fitted by \code{\link[=lmm_model]{lmm_model()}}.}

\item{pvalue}{Threshold for the p-value of outlier observations based on their Pearson residuals.}
}
\value{
A list with different elements for the diagnostics of the residuals are produced:
\itemize{
\item \code{plots}: Different plots for evaluating the normality and homocedasticity of the residuals.
\item \code{outliers}: Data frame with the identified outliers based on the Pearson residuals and the value of \code{pval}.
\item \code{Normality}: List with the results from 3 different test of the normality of the normalized residuals of the model: Shapiro - Wilk normality test,
D'Agostino normality test and Anderson - Darling normality test.
}
}
\description{
Diagnostics of residuals of the linear mixed model.
}
