% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMM_Model.R
\name{LMM_Model}
\alias{LMM_Model}
\title{Generate Linear Mixed Model for synergy calculation}
\usage{
LMM_Model(
  data,
  mouse_id = "Mouse",
  day = "Day",
  treatment = "Treatment",
  tumor_vol = "TV",
  trt_control = "Control",
  drug_a = "Drug_A",
  drug_b = "Drug_B",
  drug_ab = "Drug_AB",
  day_start = 0,
  min_observations = 1,
  show_plot = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame with the tumor growth data, in long format.
It must contain at least contain the following columns: mice IDs, days of follow-up (numeric number), treatment and tumor volume (numeric number).}

\item{mouse_id}{String indicating the name of the column with the mice IDs.}

\item{day}{String indicating the name of the column with the days of follow-up.}

\item{treatment}{String indicating the name of the column with the treatment corresponding to each mouse.}

\item{tumor_vol}{String indicating the name of the column with the tumor volume (or any other measurement representing the tumor growth).}

\item{trt_control}{String indicating the name assigned to the 'Control' group.}

\item{drug_a}{String indicating the name assigned to the 'Drug A' group.}

\item{drug_b}{String indicating the name assigned to the 'Drug B' group.}

\item{drug_ab}{String indicating the name assigned to the Combination ('Drug A' + 'Drug B') group.}

\item{day_start}{Numeric value indicating the day at which the treatment started.}

\item{min_observations}{Minimum number of observation for each mouse to be included in the analysis.}

\item{show_plot}{Logical indicating if a plot for the log of the relative tumor volume (RTV) vs Day for each mouse,
and the model calculated marginal slope for each treatment, should be produced.}

\item{...}{Additional arguments to be passed to \code{\link[nlme:lme]{nlme::lme()}} function.}
}
\value{
An object of class "lme" \code{\link[nlme:lme]{nlme::lme()}} representing the linear mixed-effects model fit.
}
\description{
Generate Linear Mixed Model for synergy calculation
}
