% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmModel.R
\name{lmmModel}
\alias{lmmModel}
\title{Generate Linear Mixed Model for synergy calculation}
\usage{
lmmModel(
  data,
  sample_id = "SampleID",
  time = "Time",
  treatment = "Treatment",
  tumor_vol = "TV",
  trt_control = "Control",
  drug_a = "Drug_A",
  drug_b = "Drug_B",
  drug_ab = "Drug_AB",
  time_start = 0,
  min_observations = 1,
  show_plot = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame with the tumor growth data, in long format.
It must contain at least contain the following columns: mice IDs, time of follow-up (numeric number), treatment and tumor volume (numeric number).}

\item{sample_id}{String indicating the name of the column with the mice IDs.}

\item{time}{String indicating the name of the column with the time of follow-up.}

\item{treatment}{String indicating the name of the column with the treatment corresponding to each sample.}

\item{tumor_vol}{String indicating the name of the column with the tumor volume (or any other measurement representing the tumor growth).}

\item{trt_control}{String indicating the name assigned to the 'Control' group.}

\item{drug_a}{String indicating the name assigned to the 'Drug A' group.}

\item{drug_b}{String indicating the name assigned to the 'Drug B' group.}

\item{drug_ab}{String indicating the name assigned to the Combination ('Drug A' + 'Drug B') group.}

\item{time_start}{Numeric value indicating the time at which the treatment started.}

\item{min_observations}{Minimum number of observation for each sample to be included in the analysis.}

\item{show_plot}{Logical indicating if a plot for the log of the relative tumor volume (RTV) vs Time for each sample,
and the model calculated marginal slope for each treatment, should be produced.}

\item{...}{Additional arguments to be passed to \code{nlme::\link[nlme:lme]{lme}}.}
}
\value{
An object of class "lme" (see \code{nlme::\link[nlme:lme]{lme}} for details) representing the linear mixed-effects model fit.
}
\description{
Generate Linear Mixed Model for synergy calculation
}
