% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmModel.R
\name{lmmModel_estimates}
\alias{lmmModel_estimates}
\title{Get estimates from a linear mixed model of tumor growth data}
\usage{
lmmModel_estimates(model)
}
\arguments{
\item{model}{An object of class "lme" representing the linear mixed-effects model fitted by \code{\link[=lmmModel]{lmmModel()}}.}
}
\value{
A data frame with the estimated values for the coefficients of the tumor growth for each treatment,
the standard deviation of the random effects, and the standard deviation of the residuals of the model.
These values can be useful for the power analysis of the model using \code{\link[=APrioriPwr]{APrioriPwr()}}.
}
\description{
Get estimates from a linear mixed model of tumor growth data
}
