% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmSynergy.R
\name{lmmSynergy}
\alias{lmmSynergy}
\title{Synergy calculation using linear-mixed models.}
\usage{
lmmSynergy(
  model,
  method = "Bliss",
  min_day = 0,
  robustSE = FALSE,
  type = "CR2",
  norm_test = "shapiroTest",
  show_plot = TRUE,
  ...
)
}
\arguments{
\item{model}{An object of class "lme" representing the linear mixed-effects model fitted by \code{\link[=lmmModel]{lmmModel()}}.}

\item{method}{String indicating the method for synergy calculation. Possible methods are "Bliss", "HSA" and "RA",
corresponding to Bliss, highest single agent and response additivity, respectively.}

\item{min_day}{Minimun day from which to start calculating synergy.}

\item{robustSE}{If TRUE, uncertainty is estimated using robust standard errors
using a sandwich estimate of the variance-covariance matrix of the regression coefficient estimates provided by \code{\link[clubSandwich:vcovCR]{clubSandwich::vcovCR()}}.}

\item{type}{Character string specifying which small-sample adjustment should be used, with available options "CR0", "CR1", "CR1p", "CR1S", "CR2", or "CR3".
See "Details" section of \code{\link[clubSandwich:vcovCR]{clubSandwich::vcovCR()}} for further information.}

\item{norm_test}{If \code{method} is set to "RA", string indicating the test for checking the
normality of the hypothesis expression of synergy (see "Warning#3" in "Description" section of \code{\link[marginaleffects:hypotheses]{marginaleffects::hypotheses()}}).
Possible values are "shapiroTest", "dagoTest" or "adTest", for Shapiro-Wilk, D'Agostino and Anderson-Darling normality tests, respectively.}

\item{show_plot}{Logical indicating if a plot with the results of the synergy calculation should be generated.}

\item{...}{Additional arguments to be passed to \code{\link[marginaleffects:hypotheses]{marginaleffects::hypotheses()}}.}
}
\value{
The function returns a list with two elements:
\itemize{
\item \code{Constrasts}: List with the outputs of the (non)-linear test for the synergy null hypothesis obtained by \code{\link[marginaleffects:hypotheses]{marginaleffects::hypotheses()}} for each day.
See \code{\link[marginaleffects:hypotheses]{marginaleffects::hypotheses()}} for more details.
\item \code{Synergy}: Data frame with the synergy results, indicating the model of synergy ("Bliss", "HSA" or "RA"), the metric (combination index and synergy score),
the value of the metric estimate (with upper and lower confidence intervals) and the p-value, for each day.
}
}
\description{
Synergy calculation using linear-mixed models.
}
