% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmDiagnostics.R
\name{ObsvsPred}
\alias{ObsvsPred}
\title{Observed vs predicted values and performance}
\usage{
ObsvsPred(model, nrow = 4, ncol = 5, ...)
}
\arguments{
\item{model}{An object of class "lme" representing the linear mixed-effects model fitted by \code{\link[=lmmModel]{lmmModel()}}.}

\item{nrow}{Number of rows of the layout to organize the observed vs predicted plots.}

\item{ncol}{Number of columns of the layout to organize the observed vs predicted plots.}

\item{...}{Additional arguments to be passed to \code{\link[performance:model_performance]{performance::model_performance()}}.}
}
\value{
A layout of the observed vs predicted values for each mouse and model performance metrics.
}
\description{
Observed vs predicted values and performance
}
\details{
The function provides visual and quantitative information about the performance of the model:
\itemize{
\item A layout of the observed and predicted values of log(relative tumor volume) vs day for each mouse,
with the actual measurements, the regression line for each mouse, and the marginal, treatment-specific,
regression line for each treatment group.
\item Performance metrics of the model obtain calculated using \code{\link[performance:model_performance]{performance::model_performance()}}.
}
}
