% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plot_ObsvsPred}
\alias{plot_ObsvsPred}
\title{Visualization of observed vs predicted values by a fitted linear mixed model of tumor growth data.}
\usage{
plot_ObsvsPred(model, nrow, ncol)
}
\arguments{
\item{model}{An object of class "lme" representing the linear mixed-effects model fitted by \code{\link[=lmmModel]{lmmModel()}}.}

\item{nrow}{Number of rows of the layout to organize the observed vs predicted plots.}

\item{ncol}{Number of columns of the layout to organize the observed vs predicted plots.}
}
\description{
Visualization of observed vs predicted values by a fitted linear mixed model of tumor growth data.
}
