% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{Plot_ObsvsPred}
\alias{Plot_ObsvsPred}
\title{Visualization of observed vs predicted values by a fitted linear mixed model of tumor growth data.}
\usage{
Plot_ObsvsPred(model, nrow, ncol)
}
\arguments{
\item{model}{An object of class "lme" representing the linear mixed-effects model fitted by \code{\link[=LMM_Model]{LMM_Model()}}.}

\item{nrow}{Number of rows of the layout to organize the observed vs predicted plots.}

\item{ncol}{Number of columns of the layout to organize the observed vs predicted plots.}
}
\description{
Visualization of observed vs predicted values by a fitted linear mixed model of tumor growth data.
}
