% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plot_ranefDiagnostics}
\alias{plot_ranefDiagnostics}
\title{Visualization of random effects diagnostics for a fitted linear mixed model of tumor growth data.}
\usage{
plot_ranefDiagnostics(model)
}
\arguments{
\item{model}{An object of class "lme" representing the linear mixed-effects model fitted by \code{\link[=lmmModel]{lmmModel()}}.}
}
\value{
A list with different plots for evaluating the normality and homoscedasticity of the random effects, including:
\itemize{
\item A normal Q-Q plot of the random effects of the model.
\item A normal Q-Q plot of the residuals by mouse.
\item Boxplots of the conditional residuals by mouse.
\item Dotplots of the pearson residuals vs fitted values by mouse.
}
}
\description{
Visualization of random effects diagnostics for a fitted linear mixed model of tumor growth data.
}
