% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plot_ranefDiagnostics}
\alias{plot_ranefDiagnostics}
\title{Plots for random effects diagnostics}
\usage{
plot_ranefDiagnostics(model)
}
\arguments{
\item{model}{An object of class "lme" representing the linear mixed-effects model fitted by \code{\link[=lmmModel]{lmmModel()}}.}
}
\value{
A list with different plots for evaluating the normality and homoscedasticity of the random effects, including:
\itemize{
\item A normal Q-Q plot of the random effects of the model.
\item A normal Q-Q plot of the residuals by sample.
\item Boxplots of the conditional residuals by sample.
\item Dotplots of the pearson residuals vs fitted values by sample.
}
}
\description{
Visualization of random effects diagnostics for a fitted linear mixed model of tumor growth data.
}
\examples{
data(grwth_data)
# Fit the model
lmm <- lmmModel(
  data = grwth_data,
  sample_id = "subject",
  time = "Time",
  treatment = "Treatment",
  tumor_vol = "TumorVolume",
  trt_control = "Control",
  drug_a = "DrugA",
  drug_b = "DrugB",
  combination = "Combination",
  show_plot = FALSE
  )
# Generate plots 
plot_ranefDiagnostics(lmm)
# Access to specific plots
plot_ranefDiagnostics(lmm)[[1]]
plot_ranefDiagnostics(lmm)[[2]]
}
