% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plot_residDiagnostics}
\alias{plot_residDiagnostics}
\title{Visualization of residuals diagnostics for a fitted linear mixed model of tumor growth data.}
\usage{
plot_residDiagnostics(model)
}
\arguments{
\item{model}{An object of class "lme" representing the linear mixed-effects model fitted by \code{\link[=lmmModel]{lmmModel()}}.}
}
\value{
A list with different plots for evaluating the normality and homoscedasticity of the random effects, including:
\itemize{
\item A normal Q-Q plot of the normalized residuals of the model.
\item A normal Q-Q plot of the normalized residuals of the model by Day.
\item A normal Q-Q plot of the normalized residuals of the model by Treatment.
\item A dotplot of pearson residuals vs fitted values.
\item A dotplot of the pearson residuals by Day and Treatment.
}
}
\description{
Visualization of residuals diagnostics for a fitted linear mixed model of tumor growth data.
}
