% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmPower.R
\name{PostHocPwr}
\alias{PostHocPwr}
\title{Post hoc power calculation based on simulations of the synergy evaluation using LMM.}
\usage{
PostHocPwr(model, nsim = 1000, method = "Bliss", pvalue = 0.05, ...)
}
\arguments{
\item{model}{An object of class "lme" representing the linear mixed-effects model fitted by \code{\link[=lmmModel]{lmmModel()}}.}

\item{nsim}{Number of simulations to perform.}

\item{method}{String indicating the method for synergy calculation. Possible methods are "Bliss", "HSA" and "RA",
corresponding to Bliss, highest single agent and response additivity, respectively.}

\item{pvalue}{Threshold for the p-value of synergy calculation to be considered statistically significant.}

\item{...}{Additional parameters to be passed to \link[nlmeU:simulateY]{nlmeU::simulateY}.}
}
\value{
Returns a numeric value of the power for the synergy calculation for the model using the method specified in \code{method}.
The power is expressed as the proportion of simulations that provides a p-value below the threshold specified in \code{pvalue}.
}
\description{
Post hoc power calculation based on simulations of the synergy evaluation using LMM.
}
