% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmPower.R
\name{Pwr_lmm}
\alias{Pwr_lmm}
\title{A priori power calculation for a hypothetical study of synergy evaluation using LMM.}
\usage{
Pwr_lmm(
  npg = 5,
  Day = c(0, 3, 5, 10),
  grwrControl,
  grwrA,
  grwrB,
  grwrComb,
  sd_ranef,
  sgma,
  sd_eval = NULL,
  sgma_eval = NULL,
  grwrComb_eval = NULL,
  method = "Bliss",
  ...
)
}
\arguments{
\item{npg}{Number of mouse per group.}

\item{Day}{Vector with the days at which the tumor volume measurements have been performed.}

\item{grwrControl}{Coefficient for Control treatment group tumor growth rate.}

\item{grwrA}{Coefficient for Drug A treatment group tumor growth rate.}

\item{grwrB}{Coefficient for Drug B treatment group tumor growth rate.}

\item{grwrComb}{Coefficient for Combination (Drug A + Drug B) treatment group tumor growth rate.}

\item{sd_ranef}{Random effects standard deviation for the model.}

\item{sgma}{Residuals standard deviation for the model.}

\item{sd_eval}{A vector with values representing the standard deviations of random effects,
through which the power for synergy calculation will be evaluated.}

\item{sgma_eval}{A vector with values representing the standard deviations of the residuals,
through which the power for synergy calculation will be evaluated.}

\item{grwrComb_eval}{A vector with values representing the coefficients for Combination treatment group tumor growth rate,
through which the power for synergy calculation will be evaluated.}

\item{method}{String indicating the method for synergy calculation. Possible methods are "Bliss" and "HSA",
corresponding to Bliss and highest single agent, respectively.}

\item{...}{Additional parameters to be passed to \link[nlmeU:Pwr.lme]{nlmeU::Pwr.lme} method.}
}
\value{
The functions returns several plots:
\itemize{
\item A plot representing the hypothetical data, with the regression lines for each
treatment group according to \code{grwrControl}, \code{grwrA}, \code{grwrB} and \code{grwrComb} values. The values
assigned to \code{sd_ranef} and \code{sgma} are also shown.
\item A plot showing the values of the power calculation depending on the values assigned to
\code{sd_eval} and \code{sgma_eval},
\item A plot showing the values of the power calculation depending on the values assigned to
\code{grwrComb_eval}.
If saved to a variable, the function saves the exemplary data frame built for the hypothetical study.
}
}
\description{
A priori power calculation for a hypothetical study of synergy evaluation using LMM.
}
