% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmDiagnostics.R
\name{ranefDiagnostics}
\alias{ranefDiagnostics}
\title{Diagnostics of random effects of the linear mixed model}
\usage{
ranefDiagnostics(model)
}
\arguments{
\item{model}{An object of class "lme" representing the linear mixed-effects model fitted by \code{\link[=lmmModel]{lmmModel()}}.}
}
\value{
A list with different elements for the diagnostics of the random effects are produced:
\itemize{
\item \code{plots}: Different plots for evaluating the normality and homoscedasticity of the random effects are produced.
\item \code{Normality}: List with the results from 3 different test of the normality of the random effects: Shapiro - Wilk normality test,
D'Agostino normality test and Anderson - Darling normality test.
\item \code{Levene.test}: results from Levene homocedasticity test of the conditional, marginal and normalized residuals by SampleID (i.e. by subject).
\item \code{Fligner.test}: results from Fligner homocedasticity test of the conditional, marginal and normalized residuals by SampleID (i.e. by subject).
}
}
\description{
\code{ranefDiagnostics} provides several plots as well as statistical test for the examination
of the normality of the random effects of the input model.
}
\details{
One of the assumptions of the model obtained with \code{\link[=lmmModel]{lmmModel()}} (as in any other linear mixed model) is that
the random effects are normally distributed:

\deqn{b_i = N(0,\psi)}

For the evaluation of this assumption, \code{ranefDiagnostics} provides Q-Q plots of the normalized residuals,
together with statistical assessment of their normality using Shapiro-Wilk, D'Agostini and Anderson-Darling normality tests.
Additionally, Q-Q plots of the normalized residuals by time point and treatment group are provided to be able to detect time
points or treatment groups which could be notably different from the others and be affecting the adequacy of the model.
Additionally, scatter plots of the conditional Pearson residuals versus fitted values and Pearson residuals per time and per
treatment are provided to give information about variability of the residuals and possible outlier observations.
Observations with absolute standardized (normalized) residuals greater than the \eqn{1-0.05/2} quantile of the standard normal distribution
are identified and reported as potential outlier observations.
}
\examples{
# Load the example data
data(grwth_data)
# Fit the model
lmm <- lmmModel(
  data = grwth_data,
  sample_id = "subject",
  time = "Time",
  treatment = "Treatment",
  tumor_vol = "TumorVolume",
  trt_control = "Control",
  drug_a = "DrugA",
  drug_b = "DrugB",
  combination = "Combination"
  )
# Run random effects diagnostics
ranef_diag <- ranefDiagnostics(lmm)

#Access to individual plots

ranef_diag$Plots[1]
ranef_diag$Plots[2]

# Access to normality tests

ranef_diag$Normality

ranef_diag$Normality$Shapiro.test

# Access to homoscedasticity tests of residuals by subject

ranef_diag$Levene.test

ranef_diag$Fligner.test

}
\references{
\itemize{
\item Pinheiro JC, Bates DM (2000). \emph{Mixed-Effects Models in S and S-PLUS}. Springer, New York. doi:10.1007/b98882 \url{https://doi.org/10.1007/b98882}.
\item Andrzej Galecki & Tomasz Burzykowski (2013) \emph{Linear Mixed-Effects Models Using R: A Step-by-Step Approach} First Edition. Springer, New York. ISBN 978-1-4614-3899-1
}
}
